requirements.txt
python-telegram-bot
python-dotenv
transformers==4.31.0
accelerate==0.21.0
einops==0.6.1
langchain==0.0.240
xformers==0.0.20
huggingface-hub
scipy
